Fixed various bugs
