+++ data_handler/cds/__init__.py
@@ -1,1 +1,1 @@
-from data_handler.cds import ClimateDataStorageHandler
+from data_handler.cds.cds import ClimateDataStorageHandler
