--- utils/__init__.py
@@ -1,1 +1,0 @@
-from utils import Utilities
