+++ utils/utils.py
@@ -1,9 +1,10 @@
 import yaml 
+import json 
 
 class Utilities():
     
     @staticmethod
-    def load_config_file(self, file_path: str) -> dict:
+    def load_config_file(file_path: str) -> dict:
         """
         Load YAML file.
 
@@ -16,3 +17,34 @@ class Utilities():
         with open(file_path, 'r') as file:
             config = yaml.safe_load(file)
         return config
+
+    @staticmethod
+    def cleaned_dict_output(response):
+        """
+        Cleans a JSON string by removing any content before the first `{` and after the last `}`.
+
+        Args:
+        response (str): The raw JSON response string.
+
+        Returns:
+        dict: A dictionary parsed from the cleaned JSON string.
+        """
+        # Find the first '{' and last '}'
+        start_idx = response.find('{')
+        end_idx = response.rfind('}')
+        
+        if start_idx == -1 or end_idx == -1:
+            print("No JSON object found in the response.")
+            return None
+        
+        # Extract the JSON string
+        json_str = response[start_idx:end_idx + 1]
+        
+        # Replace single quotes with double quotes for valid JSON
+        json_str = json_str.replace("'", '"')
+        
+        try:
+            return json.loads(json_str)
+        except json.JSONDecodeError as e:
+            print(f"Error decoding JSON: {e}")
+            return None
